requirements.txt
tkinterdnd2
pyyaml